//handling existing file

+++ src/main/java/CommanderComponents/FileCommanderOperations.java
@@ -5,10 +5,13 @@ import Adapters.FileSystemObject;
 import org.apache.commons.io.FileUtils;
 
 import javax.swing.*;
+import javax.swing.text.BadLocationException;
+import javax.swing.text.Document;
+import javax.swing.text.html.HTMLEditorKit;
+import javax.swing.text.rtf.RTFEditorKit;
 import java.awt.*;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.IOException;
+import java.io.*;
+import java.nio.charset.StandardCharsets;
 
 public class FileCommanderOperations {
     private FileCommanderFrame frame;
@@ -25,20 +28,21 @@ public class FileCommanderOperations {
     void createNewFile(String path){
         try {
             FileSystemObject file = new FileSystemObject(path);
-            if(file.getExtension().equals("-dir")){
+            if (file.getExtension().equals("-dir")) {
                 createNewFolder(path);
                 return;
             }
-            boolean isCreated = file.createNewFile();
-            System.out.println(file.createNewFile());
-            if(isCreated){
-                JOptionPane.showMessageDialog(frame,"Can't create new file","Error",1);
+            if (!handleExistingFile(path)) {
+                boolean isCreated = file.createNewFile();
+                System.out.println(file.createNewFile());
+                if (isCreated) {
+                    JOptionPane.showMessageDialog(frame, "Can't create new file", "Error", 1);
+                }
+                refreshLists();
             }
-            refreshLists();
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+            } catch(IOException e){
+                e.printStackTrace();
+            }
 
     }
     public void refreshLists(){
@@ -82,6 +86,7 @@ public class FileCommanderOperations {
         }
     }
     void createNewFolder(String path){
+        if(handleExistingFile(path))return;
         try{
             FileSystemObject file = new FileSystemObject(path);
             FileUtils.forceMkdir(file);
@@ -104,6 +109,7 @@ public class FileCommanderOperations {
         try {
             FileSystemObject fileFrom = new FileSystemObject(from);
             FileSystemObject fileTo = new FileSystemObject(to + "\\" + fileFrom.getName());
+            if(handleExistingFile(fileTo.toString()))return;
             if (!fileFrom.isDirectory()) FileUtils.copyFile(fileFrom, fileTo);
             else FileUtils.copyDirectory(fileFrom,fileTo);
         }
@@ -128,6 +134,7 @@ public class FileCommanderOperations {
     }
 
     void removeFile(String from, String to){
+        if(handleExistingFile(to))return;
         try{
             copyFile(from,to);
             deleteFile(from);
@@ -139,7 +146,9 @@ public class FileCommanderOperations {
     void renameFile(String path, String name){
         try{
             FileSystemObject file = new FileSystemObject(path);
-            boolean isRenamed = file.renameTo(new FileSystemObject(file.getParent()+"\\"+name));
+            String newPath = file.getParent()+"\\"+name;
+            if(handleExistingFile(newPath))return;
+            boolean isRenamed = file.renameTo(new FileSystemObject(newPath));
             if(!isRenamed){
                 JOptionPane.showMessageDialog(frame,"Can't rename this file");
             }
@@ -157,6 +166,112 @@ public class FileCommanderOperations {
             Desktop.getDesktop().open(file);
         }
         catch (Exception e){
+            e.printStackTrace();
+        }
+    }
+    private String readLine(InputStream is) throws IOException{
+        StringBuilder sb = new StringBuilder();
+        char t;
+        do {
+            t = (char)is.read();
+            sb.append(t);
+        }while(is.available()>0&&t!='\n');
+        return new String(sb);
+    }
+    public void writeLine(OutputStream os,String line){
+        for(int i=0;i<line.length();i++){
+            try{
+                os.write((byte)line.charAt(i));
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+    }
+    void copyFileWithoutRepeatingLines(String from, String to)  {
+
+        FileSystemObject fromFile = new FileSystemObject(from);
+        FileSystemObject toFile = new FileSystemObject(to+"\\"+fromFile.getName());
+        if(handleExistingFile(toFile.toString()))return;
+        try {
+            toFile.createNewFile();
+            InputStream inputStream =FileUtils.openInputStream(fromFile);
+            OutputStream outputStream =  FileUtils.openOutputStream(toFile);
+            String prevLine = null;
+            while(inputStream.available()>0){
+                String line = readLine(inputStream);
+                if(prevLine == null || !line.equals(prevLine)){
+                    writeLine(outputStream,line);
+                }
+                prevLine=line;
+            }
+
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+    private static String convertToRTF(String htmlStr) {
+
+        OutputStream os = new ByteArrayOutputStream();
+        HTMLEditorKit htmlEditorKit = new HTMLEditorKit();
+        RTFEditorKit rtfEditorKit = new RTFEditorKit();
+        String rtfStr = null;
+
+        htmlStr = htmlStr.replaceAll("<br.*?>","#NEW_LINE#");
+        htmlStr = htmlStr.replaceAll("</p>","#NEW_LINE#");
+        htmlStr = htmlStr.replaceAll("<p.*?>","");
+        InputStream is = new ByteArrayInputStream(htmlStr.getBytes(StandardCharsets.UTF_16));
+        try {
+            Document doc = htmlEditorKit.createDefaultDocument();
+            htmlEditorKit.read(is, doc, 0);
+            rtfEditorKit.write(os, doc, 0, doc.getLength());
+            rtfStr = os.toString();
+        } catch (IOException e) {
+            e.printStackTrace();
+        } catch (BadLocationException e) {
+            e.printStackTrace();
+        }
+        return rtfStr;
+    }
+    private boolean handleExistingFile(String path){
+        FileSystemObject file = new FileSystemObject(path);
+        if(file.exists()){
+            JOptionPane.showMessageDialog(frame,"File already exists","Error",1);
+            return true;
+        }
+        return false;
+    }
+    public void convertHtmlToRtf(String path){
+        FileSystemObject htmlFile = new FileSystemObject(path);
+        String rtfFileName = path.substring(0,path.length()-4)+"rtf";
+        System.out.println(rtfFileName);
+        FileSystemObject rtfFile = new FileSystemObject(rtfFileName);
+        if(!rtfFile.exists()){
+            try{
+                rtfFile.createNewFile();
+            }
+            catch(IOException e){
+                e.printStackTrace();
+            }
+        }
+        StringBuilder sb = new StringBuilder();
+        try {
+            InputStream is = FileUtils.openInputStream(htmlFile);
+            while(is.available()>0){
+                sb.append((char)is.read());
+            }
+            is.close();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        System.out.println(new String(sb));
+        String rtfString = convertToRTF(new String(sb));
+        try{
+            OutputStream os = FileUtils.openOutputStream(rtfFile);
+            for(int i=0;i<rtfString.length();i++){
+                os.write((byte)rtfString.charAt(i));
+            }
+            os.close();
+        } catch (IOException e) {
             e.printStackTrace();
         }
     }
